Entire file analysed from start to finish to better comment code.
Methodological improvement to SISS df to calculate the number of SISS by distance greater than 7.5km. Previously this head() function was hardcoded as 20 which was the number greater than 7.5km. By using this approach, of counting the schools rather than hard coding the 20, it makes the code more replicable and allows it to be used in future, or previous years, when the number will not always be 20